[Refactor 답변,댓글]

+++ server/src/main/java/com/codinghaezo/stackOverFlow/answer/AnswerService.java
@@ -1,5 +1,9 @@
 package com.codinghaezo.stackOverFlow.answer;
 
+import com.codinghaezo.stackOverFlow.exception.BusinessLogicException;
+import com.codinghaezo.stackOverFlow.exception.ExceptionCode;
+import com.codinghaezo.stackOverFlow.member.Member;
+import com.codinghaezo.stackOverFlow.member.MemberRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -7,6 +11,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -14,65 +19,71 @@ import java.util.Optional;
 @Transactional
 public class AnswerService {
     private final AnswerRepository answerRepository;
+    private final MemberRepository memberRepository;
 
-
-    public AnswerService(AnswerRepository answerRepository) {
+    public AnswerService(AnswerRepository answerRepository,
+                         MemberRepository memberRepository) {
         this.answerRepository = answerRepository;
+        this.memberRepository = memberRepository;
     }
 
-    @Transactional(readOnly = true)
     public Answer findAnswer(long answerId) {
         return answerRepository.findById(answerId).orElseThrow();
     }
 
-    @Transactional(readOnly = true)
-    public List<Answer> findAnswers(long Id) {
-        return answerRepository.findByQuestionId(Id);
+    public List<AnswerDto.AnswerResponseDTOV1> getAnswersByQuestionId(long questionId) {
+        List<Answer> answers = answerRepository.findByQuestionId(questionId);
+        List<AnswerDto.AnswerResponseDTOV1> responseDTOList = new ArrayList<>();
+
+        for (Answer answer : answers) {
+            AnswerDto.AnswerResponseDTOV1 responseDTO = new AnswerDto.AnswerResponseDTOV1();
+            responseDTO.setQuestionId(answer.getQuestion().getId());
+            responseDTO.setAnswerId(answer.getAnswerId());
+            responseDTO.setContent(answer.getContent());
+            responseDTO.setUserEmail(answer.getMember().getEmail());
+            responseDTO.setCreatedAt(answer.getCreatedAt());
+            responseDTO.setUpdatedAt(answer.getModifiedAt());
+            responseDTOList.add(responseDTO);
+        }
+
+        return responseDTOList;
     }
 
+    public List<AnswerDto.AnswerResponseDTOV2> findByUserEmail(String userEmail) {
+        List<Answer> answers = answerRepository.findByMemberEmail(userEmail);
+        List<AnswerDto.AnswerResponseDTOV2> responseDTOList = new ArrayList<>();
 
-   /* public Page<Answer> findAll(int page, int size) {
-        return answerRepository.findAll(PageRequest.of(page - 1, size, Sort.by("answerId").descending()));
-    } */
-
-    /*   여기도 특정사용자가 작성한 답변 전체조회하는 부분
-          public List<Answer> findAnswersByMemberIdAndQuestionId(Long memberId, Long questionId) {
-          return answerRepository.findByMember_IdAndQuestion_Id(memberId, questionId);
-      }
-     */
+            for(Answer answer : answers){
+                AnswerDto.AnswerResponseDTOV2 responseDTO = new AnswerDto.AnswerResponseDTOV2();
+                responseDTO.setQuestionTitle(answer.getQuestion().getTitle());
+                responseDTO.setQuestionId(answer.getQuestion().getId());
+                responseDTO.setContent(answer.getContent());
+                responseDTOList.add(responseDTO);
+            }
+              return responseDTOList;
+    }
     public Answer createAnswer(Answer answer) {
         return answerRepository.save(answer);
     }
 
-    public void deleteAnswer(long answerId) {
+    public void deleteAnswer(long answerId, String userEmail) {
         Answer foundAnswer = findAnswer(answerId);
+        verifyAuthor(userEmail, foundAnswer);
         answerRepository.delete(foundAnswer);
     }
 
-    public Answer updateAnswer(Answer answer, long answerId, long memberId) {
-        Answer foundAnswer = findAnswer(answerId); /*
-         1. 답변아이디로 답변글을 찾는다.
-         2. 답변글로 작성자 멤버 아이디를 찾는다. findMemberId
-             /*
-           식별 if(memberId==findMemberId)
-           else 에러 던지기(유효하지 않은 회원입니다.)
-            */
-        foundAnswer.setContent(answer.getContent()); // 찾은 답변글의 내용을 수정한다.
-        return answerRepository.save(foundAnswer); //리포지토리에 저장한다.
-    }
-
-
-    public Answer updateAnswer(Answer answer, long answerId, String membername) {
+    public Answer updateAnswer(Answer answer, long answerId, String userEmail) {
         Answer foundAnswer = findAnswer(answerId);
-        //membername 이랑 이름이랑 비교해야함 수정해야함
-        if (membername == foundAnswer.getContent()) {
+        verifyAuthor(userEmail, foundAnswer);
             foundAnswer.setContent(answer.getContent());
             return answerRepository.save(foundAnswer);
-        } else {
-            return null;
+    }
+
+    private void verifyAuthor(String principalEmail, Answer answer) {
+        String authorEmail = answer.getMember().getEmail();;
+        if (!principalEmail.equals(authorEmail)) {
+            throw new BusinessLogicException(ExceptionCode.CANNOT_CHANGE_ANSWER);
         }
-
-
     }
 
 }
